fix(ui): distinguish terminal clients in presence cards
Preserve the accepted client type in presence and restore it with the connection mode after cache eviction. Display Terminal for TUI connections while retaining App for native clients.
