+++ src/gateway/server/ws-connection/connect-session.ts
@@ -596,6 +596,7 @@ export async function attachAuthenticatedGatewayConnect(
     const authenticatedPresenceUser = currentAuthenticatedPresenceUser();
     upsertPresence(presenceKey, {
       host: connectParams.client.displayName ?? connectParams.client.id ?? os.hostname(),
+      clientId: connectParams.client.id,
       ip: isLocalClient ? undefined : reportedClientIp,
       version: connectParams.client.version,
       platform: connectParams.client.platform,
